update prod ready example
