+++ examples/sandbox/src/pages/elasticsearch-production-ready/index.jsx
@@ -4,25 +4,9 @@ import "@elastic/eui/dist/eui_theme_light.css";
 import { ApiProxyConnector } from "@elastic/search-ui-elasticsearch-connector";
 import moment from "moment";
 
-import {
-  ErrorBoundary,
-  Facet,
-  SearchProvider,
-  SearchBox,
-  Results,
-  PagingInfo,
-  ResultsPerPage,
-  Paging,
-  Sorting,
-  WithSearch
-} from "@elastic/react-search-ui";
-import {
-  BooleanFacet,
-  Layout,
-  SingleLinksFacet,
-  SingleSelectFacet
-} from "@elastic/react-search-ui-views";
+import { SearchProvider } from "@elastic/react-search-ui";
 import "@elastic/react-search-ui-views/lib/styles/styles.css";
+import { ElasticsearchProdReady } from "./elasticsearchProdReady";
 
 const connector = new ApiProxyConnector({
   basePath: "http://localhost:3001/api"
@@ -154,150 +138,10 @@ const config = {
   }
 };
 
-const SORT_OPTIONS = [
-  {
-    name: "Relevance",
-    value: []
-  },
-  {
-    name: "Title",
-    value: [
-      {
-        field: "title.keyword",
-        direction: "asc"
-      }
-    ]
-  },
-  {
-    name: "State",
-    value: [
-      {
-        field: "states.keyword",
-        direction: "asc"
-      }
-    ]
-  },
-  {
-    name: "State -> Title",
-    value: [
-      {
-        field: "states.keyword",
-        direction: "asc"
-      },
-      {
-        field: "title.keyword",
-        direction: "asc"
-      }
-    ]
-  },
-  {
-    name: "Heritage Site -> State -> Title",
-    value: [
-      {
-        field: "world_heritage_site.keyword",
-        direction: "asc"
-      },
-      {
-        field: "states.keyword",
-        direction: "asc"
-      },
-      {
-        field: "title.keyword",
-        direction: "asc"
-      }
-    ]
-  }
-];
-
 export default function App() {
   return (
     <SearchProvider config={config}>
-      <WithSearch
-        mapContextToProps={({ wasSearched }) => ({
-          wasSearched
-        })}
-      >
-        {({ wasSearched }) => {
-          return (
-            <div className="App">
-              <ErrorBoundary>
-                <Layout
-                  header={
-                    <SearchBox
-                      autocompleteMinimumCharacters={3}
-                      autocompleteResults={{
-                        linkTarget: "_blank",
-                        sectionTitle: "Results",
-                        titleField: "title",
-                        urlField: "nps_link",
-                        shouldTrackClickThrough: true,
-                        clickThroughTags: ["test"]
-                      }}
-                      autocompleteSuggestions={true}
-                      debounceLength={0}
-                    />
-                  }
-                  sideContent={
-                    <div>
-                      {wasSearched && (
-                        <Sorting label={"Sort by"} sortOptions={SORT_OPTIONS} />
-                      )}
-                      <Facet
-                        field="states.keyword"
-                        label="States"
-                        filterType="any"
-                        isFilterable={true}
-                      />
-                      <Facet
-                        field="world_heritage_site.keyword"
-                        label="World Heritage Site"
-                        view={BooleanFacet}
-                      />
-                      <Facet
-                        field="visitors"
-                        label="Visitors"
-                        view={SingleLinksFacet}
-                      />
-                      <Facet
-                        field="date_established"
-                        label="Date Established"
-                        isFilterable={true}
-                        filterType="any"
-                      />
-                      <Facet
-                        field="location"
-                        label="Distance"
-                        filterType="any"
-                      />
-                      <Facet field="visitors" label="visitors" />
-                      <Facet
-                        field="acres"
-                        label="Acres"
-                        view={SingleSelectFacet}
-                      />
-                    </div>
-                  }
-                  bodyContent={
-                    <Results
-                      titleField="title"
-                      urlField="nps_link"
-                      thumbnailField="image_url"
-                      shouldTrackClickThrough={true}
-                    />
-                  }
-                  bodyHeader={
-                    <React.Fragment>
-                      {wasSearched && <PagingInfo />}
-                      {wasSearched && <ResultsPerPage />}
-                    </React.Fragment>
-                  }
-                  bodyFooter={<Paging />}
-                />
-              </ErrorBoundary>
-            </div>
-          );
-        }}
-      </WithSearch>
+      <ElasticsearchProdReady />
     </SearchProvider>
   );
 }
